// bridgebio/bridgebio-backend/src/common/types/permissions.ts
export enum Permissions{
    SUPER_ADMIN = 'SUPER_ADMIN',
    MEDS = 'MEDS',
    EVIDENCE_GENERATION_FRAMEWORK = 'EVIDENCE_GENERATION_FRAMEWORK',
    EVIDENCE_GENERATION_PRIORITIES = 'EVIDENCE_GENERATION_PRIORITIES',
    RESEARCH_INITIATIVES = 'RESEARCH_INITIATIVES',
    RESEARCH_TEAMS = 'RESEARCH_TEAMS',
    EVIDENCE_LITERATURE_LIBRARY = 'EVIDENCE_LITERATURE_LIBRARY',
}